requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
psycopg==3.1.8
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
